Commit - Pakage ec.edu.ups.model
Corrección del modelo.

+++ src/ec/edu/ups/model/Privilege.java
@@ -19,9 +19,9 @@ public class Privilege {
     
     private int priId;
     private String priName;
-    private double priDiscountPct;
+    private int priDiscountPct;
     
-    public Privilege(int priId, String priName, double priDiscountPot) {
+    public Privilege(int priId, String priName, int priDiscountPot) {
         this.priId = priId;
         this.priName = priName;
         this.priDiscountPct = priDiscountPot;
@@ -43,49 +43,14 @@ public class Privilege {
         this.priName = priName;
     }
 
-    public double getPriDiscountPot() {
+    public int getPriDiscountPot() {
         return priDiscountPct;
     }
 
-    public void setPriDiscountPot(double priDiscountPct) {
+    public void setPriDiscountPot(int priDiscountPct) {
         this.priDiscountPct = priDiscountPct;
     }
-//
-//    public List<User> getPriUser() {
-//        return priUsers;
-//    }
-//
-//    public void setPriUsers(List<User> priUsers) {
-//        this.priUsers = priUsers;
-//    }
-//
-//    public boolean createUser(User user){
-//        if (this.priUsers == null)
-//            this.priUsers = new ArrayList<>();
-//        this.priUsers.add(user);
-//        return true;
-//    }
-//    
-//    public User readUser(int useId){
-//        if (this.priUsers == null)
-//            throw new NullPointerException();
-//        return this.priUsers.get(useId);
-//    }
-//    
-//    public boolean updateUser(int useId, User user){
-//        if (this.priUsers == null)
-//            throw new NullPointerException();
-//        this.priUsers.set(useId, user);
-//        return true;
-//    }
-//    
-//    public boolean deleteUser(int useId){
-//        if (this.priUsers == null)
-//            throw new NullPointerException();
-//        this.priUsers.remove(useId);
-//        return true;
-//    }
-
+    
     @Override
     public String toString() {
         return "Privilege{" + "priId=" + priId + ", priName=" + priName
